+++ projeto/budget-parent/budget-business/src/main/java/com/dupont/budget/service/impl/ReportServiceBean.java
@@ -24,32 +24,32 @@ import com.dupont.budget.service.ReportService;
 
 /**
  * Implementação dos serviços que fornecem os dados de relatorio.
- * 
+ *
  * @author <a href="mailto:asouza@redhat.com">Ângelo Galvão</a>
  * @since 2015
  *
  */
 @Stateless
 public class ReportServiceBean implements ReportService {
-	
+
 	@PersistenceContext(unitName="budget-pu")
 	private EntityManager em;
 
 	@Override
 	public byte[] exportDatabase() {
-		
-		List<SolicitacaoPagamento> solicitacoes = em.createNamedQuery("SolicitacaoPagamento.findAll", SolicitacaoPagamento.class).getResultList();		
-		List<Budget> budgets                    = em.createNamedQuery("Budget.findAll"              , Budget.class).getResultList();		
+
+		List<SolicitacaoPagamento> solicitacoes = em.createNamedQuery("SolicitacaoPagamento.findAll", SolicitacaoPagamento.class).getResultList();
+		List<Budget> budgets                    = em.createNamedQuery("Budget.findAll"              , Budget.class).getResultList();
 		List<Forecast> forecast                 = em.createNamedQuery("Forecast.findAll"            , Forecast.class).getResultList();
-		
+
 		Workbook wb = new HSSFWorkbook();
 		Sheet sheet = wb.createSheet();
-				
+
 		int rowNumber = 0;
-		
-		// Criar Cabeçalho de solicitacao de pagamento 
+
+		// Criar Cabeçalho de solicitacao de pagamento
 		Row row   = sheet.createRow(rowNumber++);
-		
+
 		row.createCell(0).setCellValue("Responsável pelo Envio da Solicitação");
 		row.createCell(1).setCellValue("Fornecedor");
 		row.createCell(2).setCellValue("Nr da Nota");
@@ -62,17 +62,17 @@ public class ReportServiceBean implements ReportService {
 		row.createCell(9).setCellValue("Distrito");
 		row.createCell(10).setCellValue("ERC/CCE");
 		row.createCell(11).setCellValue("Valor");
-				
-		
-		for (SolicitacaoPagamento solicitacao : solicitacoes) {		
-			
+
+
+		for (SolicitacaoPagamento solicitacao : solicitacoes) {
+
 			Set<DespesaSolicitacaoPagamento> despesas = solicitacao.getDespesas();
-			
+
 			if( despesas == null || despesas.isEmpty() )
 				continue;
-			
+
 			row = sheet.createRow(rowNumber++);
-			
+
 			for (DespesaSolicitacaoPagamento despesa : despesas) {
 				row.createCell(0).setCellValue(solicitacao.getUsuarioCriador().getNome()); // Responsavel
 				row.createCell(1).setCellValue(solicitacao.getFornecedor().getNome()); // Fornecedor
@@ -85,17 +85,17 @@ public class ReportServiceBean implements ReportService {
 				row.createCell(8).setCellValue(despesa.getCultura().getNome()); // cultura
 				row.createCell(9).setCellValue(despesa.getDistrito().getNome()); // distrito
 				row.createCell(10).setCellValue(despesa.getVendedor()==null? "":despesa.getVendedor().getNome()); // ERC/CCE
-				row.createCell(11).setCellValue(despesa.getValor()); // valor
+				row.createCell(11).setCellValue(despesa.getValor().doubleValue()); // valor
 			}
-			
-			
+
+
 		}
-		
-		
+
+
 		// Criar cabeçalho de Budget
 		rowNumber += 2;
 		row = sheet.createRow(rowNumber);
-		
+
 		row.createCell(0).setCellValue("Area");
 		row.createCell(1).setCellValue("Responsável pelo dentro de custo");
 		row.createCell(2).setCellValue("Centro de custo");
@@ -111,25 +111,25 @@ public class ReportServiceBean implements ReportService {
 		row.createCell(12).setCellValue("Budget Proposto");
 		row.createCell(13).setCellValue("Budget Aprovado");
 		row.createCell(14).setCellValue("Comentário");
-		
+
 		for (Budget budget : budgets) {
-			
+
 			Set<Despesa> despesas = budget.getDespesas();
-			
+
 			if( despesas == null || despesas.isEmpty() )
 				continue;
-			
+
 			for (Despesa despesa : despesas) {
-				
+
 				row.createCell(0);
 			}
-			
+
 		}
-		
+
 		// Criar cabeçalho de Forecast
 		rowNumber += 2;
 		row = sheet.createRow(rowNumber);
-		
+
 		row.createCell(0).setCellValue("Area");
 		row.createCell(1).setCellValue("Responsável pelo centro de custo");
 		row.createCell(2).setCellValue("Centro de Custo");
@@ -144,7 +144,7 @@ public class ReportServiceBean implements ReportService {
 		row.createCell(11).setCellValue("Mês");
 		row.createCell(12).setCellValue("Valor Forecast ou Gasto Efetivo");
 		row.createCell(13).setCellValue("Valor Comprometido");
-		
+
 		// Criar resposta
 		ByteArrayOutputStream bos = new ByteArrayOutputStream();
 		try {
@@ -152,7 +152,7 @@ public class ReportServiceBean implements ReportService {
 		} catch (IOException e) {
 			throw new SystemException("ERRO ao gravar o ByteArrayOutputStream.");
 		}
-		
+
 		return bos.toByteArray();
 	}
