+++ projeto/budget-parent/budget-datamodel/src/main/java/com/dupont/budget/report/model/ReportBudgetOrcadoUtilizadoDetail.java
@@ -1,41 +1,42 @@
 package com.dupont.budget.report.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Detalhe dos relatorio do tipo BUDGET ORÇADO x UTILIZADO.
- * 
+ *
  * @author <a href="mailto:asouza@redhat.com">Ângelo Galvão</a>
  * @since 2015
  *
  */
 public class ReportBudgetOrcadoUtilizadoDetail implements Serializable {
-	
+
 	private static final long serialVersionUID = -6529967707879672477L;
-	
-	private Double orcado;
-	private Double utilizado;
+
+	private BigDecimal orcado;
+	private BigDecimal utilizado;
 	private String detail;
-	
-	public ReportBudgetOrcadoUtilizadoDetail(String detail, Double orcado, Double utilizado) {
+
+	public ReportBudgetOrcadoUtilizadoDetail(String detail, BigDecimal orcado, BigDecimal utilizado) {
 		this.detail = detail;
 		this.orcado = orcado;
 		this.utilizado = utilizado;
 	}
 
-	public Double getOrcado() {
+	public BigDecimal getOrcado() {
 		return orcado;
 	}
 
-	public void setOrcado(Double orcado) {
+	public void setOrcado(BigDecimal orcado) {
 		this.orcado = orcado;
 	}
 
-	public Double getUtilizado() {
+	public BigDecimal getUtilizado() {
 		return utilizado;
 	}
 
-	public void setUtilizado(Double utilizado) {
+	public void setUtilizado(BigDecimal utilizado) {
 		this.utilizado = utilizado;
 	}
